+++ packages/ipc/src/index.ts
@@ -1,0 +1,13 @@
+export {
+  allChannels,
+  appInfo,
+  formatIssues,
+  windowSetTitle,
+  type AppInfoRequest,
+  type AppInfoResponse,
+  type Channel,
+  type WindowSetTitleRequest,
+  type WindowSetTitleResponse,
+} from './channels'
+export { CHANNEL_NAMES, type ChannelName } from './names'
+export type { PianoBridge } from './bridge'